[exotica_core_task_maps] Fix float conversion

+++ exotations/exotica_core_task_maps/src/sphere_collision.cpp
@@ -53,7 +53,7 @@ void SphereCollision::Instantiate(const SphereCollisionInitializer& init)
     for (auto& it : groups_)
     {
         std_msgs::ColorRGBA col = RandomColor();
-        col.a = init.Alpha;
+        col.a = static_cast<float>(init.Alpha);
         for (int i : it.second)
         {
             debug_msg_.markers[i].color = col;
